requirements.txt
biopython
streamlit